--- tests/browser/account.test.js
@@ -1,21 +1,0 @@
-import { expect, describe, it } from "vitest";
-
-describe(`account management`, {}, () => {
-  describe(`the smart account is not yet deployed and you do not have funds`, {}, () => {
-    it.todo(`should provide the address and position`);
-  });
-
-  describe(`the smart account is not yet deployed but you have funds`, {}, () => {
-    it.todo(`should provide the address and position`);
-
-    it.todo(`should provide the ability to send funds`);
-  });
-
-  describe(`the smart account is deployed`, {}, () => {
-    it.todo(`should provide the address and position`);
-
-    it.todo(`should provide the ability to send funds`);
-
-    it.todo(`should provide the transaction history`);
-  });
-});
